requirements.txt
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.2
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
